expanded checking to report if either interface goes down

+++ iblinkwatch.sh
@@ -6,38 +6,43 @@
 
 #email address for notifications to be sent
 EMAIL="strosahl@jlab.org"
-MUTT='/usr/bin/mutt'
+SENDER="set from=$HOSTNAME@jlab.org"
 
 #primary interface 
-PRIMARY="ib1"
+PRIMARY="ib0"
 
 TIMELIMIT=3
 
 #active string
 ACTIVE="Currently Active Slave"
+DOWN="down"
 
 #file created to tell script that the error has already been detected
 TOUCH="/bin/touch"
-TOUCHFILE="ibalert"
+TOUCHFILE="/tmp/ibalert"
 
 INFO="/proc/net/bonding/bond0"
 
 #information on current bond status
 CURRENT=`cat $INFO | grep "$ACTIVE"`
 
+INDIVIDUAL_LINKS=`cat $INFO | grep "$DOWN"`
+
 #check to see if the current active link is the defined primary link
 #if it is, and the touchfile exists it removes the touchfule sends an
 #email notification, and exits
-if [[ $CURRENT == *"$PRIMARY"* ]]
+if [[ $CURRENT ==  *"$PRIMARY"* && $INDIVIDUAL_LINKS != *"$DOWN"* ]]
 then
 	if [[ -f $TOUCHFILE ]]
 	then
 		#status restored
-		$MUTT -s "IB link restored to $PRIMARY on $HOSTNAME" $EMAIL < $INFO
+		/usr/bin/mutt -e "$SENDER" -s "IB link restored to $PRIMARY on $HOSTNAME" $EMAIL < $INFO
 		rm $TOUCHFILE
 	fi
 	exit	
 fi
+
+echo "here"
 
 #If the touch file is older then the time limit we remove it
 if [[ -f $TOUCHFILE ]]
@@ -52,7 +57,7 @@ then
 	exit
 else
 	#at this point a notification needs to be sent
-	$MUTT -s "ALERT IB link status change on $HOSTNAME" $EMAIL < $INFO
+	/usr/bin/mutt -e "$SENDER" -s "ALERT IB link status change on $HOSTNAME" $EMAIL < $INFO
 	$TOUCH $TOUCHFILE
 	exit
 fi
